requirements.txt
langchain
langchain-community
langchain-chroma
langchain-huggingface
langgraph
sentence-transformers
torch
chromadb
langchain-text-splitters
requests
python-dotenv
tqdm
flask